# executor: skip row comparison in steam aggregation when necessary