--- a/executor/aggregate.go
+++ b/executor/aggregate.go
@@ -1298,13 +1298,15 @@ func (e *vecGroupChecker) evalGroupItemsAndResolveGroups(item expression.Express
 		vals := col.Int64s()
 		for i := 1; i < numRows; i++ {
 			isNull := col.IsNull(i)
-			switch e.sameGroup[i] {
-			case !previousIsNull && !isNull:
-				if vals[i] != vals[i-1] {
+			if e.sameGroup[i] {
+				switch {
+				case !previousIsNull && !isNull:
+					if vals[i] != vals[i-1] {
+						e.sameGroup[i] = false
+					}
+				case isNull != previousIsNull:
 					e.sameGroup[i] = false
 				}
-			case isNull != previousIsNull:
-				e.sameGroup[i] = false
 			}
 			previousIsNull = isNull
 		}
@@ -1312,13 +1314,15 @@ func (e *vecGroupChecker) evalGroupItemsAndResolveGroups(item expression.Express
 		vals := col.Float64s()
 		for i := 1; i < numRows; i++ {
 			isNull := col.IsNull(i)
-			switch e.sameGroup[i] {
-			case !previousIsNull && !isNull:
-				if vals[i] != vals[i-1] {
+			if e.sameGroup[i] {
+				switch {
+				case !previousIsNull && !isNull:
+					if vals[i] != vals[i-1] {
+						e.sameGroup[i] = false
+					}
+				case isNull != previousIsNull:
 					e.sameGroup[i] = false
 				}
-			case isNull != previousIsNull:
-				e.sameGroup[i] = false
 			}
 			previousIsNull = isNull
 		}
@@ -1326,13 +1330,15 @@ func (e *vecGroupChecker) evalGroupItemsAndResolveGroups(item expression.Express
 		vals := col.Decimals()
 		for i := 1; i < numRows; i++ {
 			isNull := col.IsNull(i)
-			switch e.sameGroup[i] {
-			case !previousIsNull && !isNull:
-				if vals[i].Compare(&vals[i-1]) != 0 {
+			if e.sameGroup[i] {
+				switch {
+				case !previousIsNull && !isNull:
+					if vals[i].Compare(&vals[i-1]) != 0 {
+						e.sameGroup[i] = false
+					}
+				case isNull != previousIsNull:
 					e.sameGroup[i] = false
 				}
-			case isNull != previousIsNull:
-				e.sameGroup[i] = false
 			}
 			previousIsNull = isNull
 		}
@@ -1340,13 +1346,15 @@ func (e *vecGroupChecker) evalGroupItemsAndResolveGroups(item expression.Express
 		vals := col.Times()
 		for i := 1; i < numRows; i++ {
 			isNull := col.IsNull(i)
-			switch e.sameGroup[i] {
-			case !previousIsNull && !isNull:
-				if vals[i].Compare(vals[i-1]) != 0 {
+			if e.sameGroup[i] {
+				switch {
+				case !previousIsNull && !isNull:
+					if vals[i].Compare(vals[i-1]) != 0 {
+						e.sameGroup[i] = false
+					}
+				case isNull != previousIsNull:
 					e.sameGroup[i] = false
 				}
-			case isNull != previousIsNull:
-				e.sameGroup[i] = false
 			}
 			previousIsNull = isNull
 		}
@@ -1354,13 +1362,15 @@ func (e *vecGroupChecker) evalGroupItemsAndResolveGroups(item expression.Express
 		vals := col.GoDurations()
 		for i := 1; i < numRows; i++ {
 			isNull := col.IsNull(i)
-			switch e.sameGroup[i] {
-			case !previousIsNull && !isNull:
-				if vals[i] != vals[i-1] {
+			if e.sameGroup[i] {
+				switch {
+				case !previousIsNull && !isNull:
+					if vals[i] != vals[i-1] {
+						e.sameGroup[i] = false
+					}
+				case isNull != previousIsNull:
 					e.sameGroup[i] = false
 				}
-			case isNull != previousIsNull:
-				e.sameGroup[i] = false
 			}
 			previousIsNull = isNull
 		}