<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

Related to: [#12903](https://github.com/pingcap/tidb/pull/12903)

Problem Summary:
When the `e.sameGroup[i] == false`  then there is no need to compare `row[i]`.

### What is changed and how it works?
When  `e.sameGroup[i] == false` , then we can skip the comparison of `row[i]`.

Benchmark for [BenchmarkAggGroupByNDV](https://github.com/pingcap/tidb/blob/a9dc72cf25589bc00260f1af4e5a17dac0f66474/executor/benchmark_test.go#L383):
```
name                                                                                                          old time/op    new time/op    delta
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:10,_hasDistinct:false,_rows:10000000,_concurrency:4)-4          129ms ±17%     114ms ± 9%  -11.48%  (p=0.012 n=10+8)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:100,_hasDistinct:false,_rows:10000000,_concurrency:4)-4         131ms ±14%     124ms ±24%     ~     (p=0.182 n=10+9)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:10000000,_concurrency:4)-4        133ms ±16%     133ms ±24%     ~     (p=0.400 n=9+10)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:10000,_hasDistinct:false,_rows:10000000,_concurrency:4)-4       203ms ± 9%     190ms ±24%   -6.30%  (p=0.046 n=8+9)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:100000,_hasDistinct:false,_rows:10000000,_concurrency:4)-4      235ms ±30%     211ms ±39%     ~     (p=0.143 n=10+10)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:1000000,_hasDistinct:false,_rows:10000000,_concurrency:4)-4     271ms ± 0%     251ms ± 4%     ~     (p=0.400 n=1+4)

name                                                                                                          old alloc/op   new alloc/op   delta
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:10,_hasDistinct:false,_rows:10000000,_concurrency:4)-4         34.1kB ± 0%    34.1kB ± 0%     ~     (p=0.880 n=9+10)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:100,_hasDistinct:false,_rows:10000000,_concurrency:4)-4        35.6kB ± 0%    35.6kB ± 0%     ~     (p=0.277 n=10+9)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:10000000,_concurrency:4)-4       70.1kB ± 0%    70.1kB ± 0%     ~     (p=0.711 n=10+9)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:10000,_hasDistinct:false,_rows:10000000,_concurrency:4)-4      70.1kB ± 0%    70.1kB ± 0%     ~     (p=0.106 n=8+9)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:100000,_hasDistinct:false,_rows:10000000,_concurrency:4)-4     45.7kB ± 0%    45.7kB ± 0%     ~     (p=0.529 n=8+9)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:1000000,_hasDistinct:false,_rows:10000000,_concurrency:4)-4    40.3kB ± 0%    40.6kB ± 2%     ~     (p=1.000 n=1+4)

name                                                                                                          old allocs/op  new allocs/op  delta
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:10,_hasDistinct:false,_rows:10000000,_concurrency:4)-4           34.0 ± 0%      34.0 ± 0%     ~     (all equal)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:100,_hasDistinct:false,_rows:10000000,_concurrency:4)-4          38.0 ± 0%      38.0 ± 0%     ~     (all equal)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:10000000,_concurrency:4)-4         50.0 ± 0%      50.0 ± 0%     ~     (all equal)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:10000,_hasDistinct:false,_rows:10000000,_concurrency:4)-4        51.0 ± 0%      51.0 ± 0%     ~     (all equal)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:100000,_hasDistinct:false,_rows:10000000,_concurrency:4)-4       51.0 ± 0%      51.0 ± 0%     ~     (all equal)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:1000000,_hasDistinct:false,_rows:10000000,_concurrency:4)-4      51.0 ± 0%      51.2 ± 1%     ~     (p=1.600 n=1+4)
```

### Related changes

- Need to cherry-pick to the release branch

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test


### Release note <!-- bugfixes or new feature need a release note -->
- skip row comparison in steam aggregation when necessary.